Finish implementing CPU

+++ GBEmu/IBus.cs
@@ -2,6 +2,6 @@
 {
     public interface IBus : IIoDevice
     {
-        void AttachDevice(uint startAddress, uint endAddress, IIoDevice device);
+        void AttachDevice(ushort startAddress, ushort endAddress, IIoDevice device);
     }
 }
